Move trial debugger to the app router
