--- apps/store/src/pages/debugger/trial.tsx
@@ -1,1 +1,0 @@
-export { DebuggerTrialPage as default } from '@/features/widget/DebuggerTrialPage'
