+++ apps/store/src/utils/useHighlightAnimation.ts
@@ -1,3 +1,5 @@
+'use client'
+
 import { type KeyboardEvent, useCallback, useRef } from 'react'
 import { theme } from 'ui'
