+++ packages/ui/src/components/Heading/Heading.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import isPropValid from '@emotion/is-prop-valid'
 import styled from '@emotion/styled'
 import React from 'react'
